refactor: remove TensorSharding module
